[fix:17] - resolving component checkboxRow
